requirements_clean.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM and AI
google-generativeai>=0.3.0

# Data processing
pandas>=2.0.0
plotly>=5.15.0

# File processing
pdfplumber>=0.7.0
python-docx>=0.8.11

# PDF generation (optional)
reportlab>=4.0.0

# API requests
requests>=2.31.0
